Don't allow rescanning and deleting/excluding while project is running

+++ src/main/java/org/nevertouchgrass/prolific/service/ProjectDeleteService.java
@@ -3,10 +3,12 @@ package org.nevertouchgrass.prolific.service;
 import lombok.RequiredArgsConstructor;
 import org.nevertouchgrass.prolific.constants.action.DeleteProjectAction;
 import org.nevertouchgrass.prolific.constants.action.ExcludeProjectAction;
+import org.nevertouchgrass.prolific.model.notification.ErrorNotification;
 import org.nevertouchgrass.prolific.model.notification.InfoNotification;
 import org.nevertouchgrass.prolific.service.notification.NotificationService;
 import org.nevertouchgrass.prolific.service.permissions.PermissionRegistry;
 import org.nevertouchgrass.prolific.service.permissions.contract.PermissionChecker;
+import org.nevertouchgrass.prolific.service.process.ProcessService;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -20,6 +22,7 @@ public class ProjectDeleteService {
     private final PermissionRegistry permissionRegistry;
     private final NotificationService notificationService;
     private final ProjectExcluderService projectExcluderService;
+    private final ProcessService processService;
 
 
     public void deleteProject(DeleteProjectAction action) {
@@ -29,6 +32,10 @@ public class ProjectDeleteService {
             PermissionChecker<DeleteProjectAction> castedChecker = (PermissionChecker<DeleteProjectAction>) permissionChecker;
 
             if (castedChecker.hasPermission(action)) {
+                if (!processService.getObservableLiveProcesses().get(action.project()).isEmpty()) {
+                    notificationService.notifyError(ErrorNotification.of(null, "You can't exclude while process is running"));
+                    return;
+                }
                 projectExcluderService.excludeProject(new ExcludeProjectAction(action.project()));
 
                 deleteDirectory(Path.of(action.project().getPath()));
